backend/requirements.txt
# SentientIQ Backend Requirements
# The truth engine dependencies

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Async
asyncio==3.4.3
aiofiles==23.2.1

# NATS JetStream
nats-py==2.6.0

# Redis
redis==5.0.1

# AWS / S3
boto3==1.34.0
botocore==1.34.0

# Data Processing
pandas==2.1.4
pyarrow==14.0.2
numpy==1.26.2
scipy==1.11.4
scikit-learn==1.3.2

# Auth / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
ruff==0.1.9